+++ src/main/java/com/habitpay/habitpay/domain/challengepost/application/ChallengePostDeleteService.java
@@ -2,7 +2,7 @@ package com.habitpay.habitpay.domain.challengepost.application;
 
 import com.habitpay.habitpay.domain.challengepost.dao.ChallengePostRepository;
 import com.habitpay.habitpay.domain.challengepost.domain.ChallengePost;
-import com.habitpay.habitpay.domain.postphoto.application.PostPhotoService;
+import com.habitpay.habitpay.domain.postphoto.application.PostPhotoDeleteService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -13,13 +13,13 @@ import org.springframework.stereotype.Service;
 public class ChallengePostDeleteService {
 
     private final ChallengePostRepository challengePostRepository;
-    private final PostPhotoService postPhotoService;
+    private final PostPhotoDeleteService postPhotoDeleteService;
     private final ChallengePostSearchService challengePostSearchService;
 
     public void delete(Long id) {
         ChallengePost challengePost = challengePostSearchService.findById(id);
 
-        postPhotoService.deleteAllByPost(challengePost);
+        postPhotoDeleteService.deleteAllByPost(challengePost);
         challengePostRepository.delete(challengePost);
     }
 }
